Move rig_state definition(s) from rig.h to include/hamlib/rig_state.h
Compiling with NO_OLD_INCLUDES still won't work until the state
  structure(s) are moved out of rig_struct, but at least I get the
  right compiler errors.
